+++ crates/solc-json-interface/src/standard_json/output/source.rs
@@ -20,7 +20,8 @@ pub struct Source {
     /// The source code ID.
     pub id: usize,
     /// The source code AST.
-    pub ast: Option<serde_json::Value>,
+    #[serde(default, skip_serializing_if = "serde_json::Value::is_null")]
+    pub ast: serde_json::Value,
 }
 
 #[cfg(feature = "resolc")]
@@ -29,7 +30,10 @@ impl Source {
     ///
     /// Is used for projects compiled without `solc`.
     pub fn new(id: usize) -> Self {
-        Self { id, ast: None }
+        Self {
+            id,
+            ast: Default::default(),
+        }
     }
 
     /// Checks the AST node for the usage of send or transfer address methods.